requirements.txt
astroid==2.12.9
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dill==0.3.5.1
dnspython==2.2.1
docopt==0.6.2
email-validator==1.3.0
Flask==2.2.2
Flask-WTF==1.0.1
google-play-scraper==1.2.2
greenlet==1.1.3
idna==3.4
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mypy==0.971
mypy-extensions==0.4.3
packaging==21.3
pipreqs==0.4.11
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pydantic==1.10.2
pylint==2.15.2
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a27
sqlmodel==0.0.8
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
yarg==0.1.9
